requirements.txt
Flask==3.0.0
SpeechRecognition
fpdf==1.7.2
standard-aifc
audioop-lts
